pyproject.toml
[project]
name = "cc-mcp"
version = "0.1.0"
description = "MCP Server for intelligent context management and long-term dialogue consistency"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.2.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.8.0",
    "python-dotenv>=1.0.0"
]
